♻️ refactor: refactor the common settings
