+++ src/app/settings/common/Common.tsx
@@ -1,26 +1,20 @@
-import { Form, type ItemGroup, SelectWithImg, SliderWithInput } from '@lobehub/ui';
-import { Form as AntForm, App, Button, Input, Select } from 'antd';
+import { Form, type ItemGroup } from '@lobehub/ui';
+import { Form as AntForm, App, Button, Input } from 'antd';
 import isEqual from 'fast-deep-equal';
-import { AppWindow, Monitor, Moon, Palette, Sun } from 'lucide-react';
+import { AppWindow } from 'lucide-react';
 import { signIn, signOut } from 'next-auth/react';
 import { memo, useCallback, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { FORM_STYLE } from '@/const/layoutTokens';
 import { DEFAULT_SETTINGS } from '@/const/settings';
-import { imageUrl } from '@/const/url';
-import AvatarWithUpload from '@/features/AvatarWithUpload';
 import { useOAuthSession } from '@/hooks/useOAuthSession';
-import { localeOptions } from '@/locales/resources';
 import { useChatStore } from '@/store/chat';
 import { useFileStore } from '@/store/file';
 import { useGlobalStore } from '@/store/global';
-import { settingsSelectors } from '@/store/global/selectors';
+import { settingsSelectors } from '@/store/global/slices/settings/selectors';
 import { useSessionStore } from '@/store/session';
 import { useToolStore } from '@/store/tool';
-import { switchLang } from '@/utils/switchLang';
-
-import { ThemeSwatchesNeutral, ThemeSwatchesPrimary } from '../features/ThemeSwatches';
 
 type SettingItemGroup = ItemGroup;
 
@@ -45,13 +39,8 @@ const Common = memo<SettingsCommonProps>(({ showAccessCodeConfig, showOAuthLogin
   ]);
   const [removeAllFiles] = useFileStore((s) => [s.removeAllFiles]);
   const removeAllPlugins = useToolStore((s) => s.removeAllPlugins);
-
   const settings = useGlobalStore(settingsSelectors.currentSettings, isEqual);
-  const [setThemeMode, setSettings, resetSettings] = useGlobalStore((s) => [
-    s.switchThemeMode,
-    s.setSettings,
-    s.resetSettings,
-  ]);
+  const [setSettings, resetSettings] = useGlobalStore((s) => [s.setSettings, s.resetSettings]);
 
   const { message, modal } = App.useApp();
 
@@ -102,107 +91,6 @@ const Common = memo<SettingsCommonProps>(({ showAccessCodeConfig, showOAuthLogin
       title: t('danger.clear.confirm'),
     });
   }, []);
-
-  const theme: SettingItemGroup = {
-    children: [
-      {
-        children: <AvatarWithUpload />,
-        label: t('settingTheme.avatar.title'),
-        minWidth: undefined,
-      },
-      {
-        children: (
-          <SelectWithImg
-            defaultValue={settings.themeMode}
-            height={60}
-            onChange={setThemeMode}
-            options={[
-              {
-                icon: Sun,
-                img: imageUrl('theme_light.webp'),
-                label: t('settingTheme.themeMode.light'),
-                value: 'light',
-              },
-              {
-                icon: Moon,
-                img: imageUrl('theme_dark.webp'),
-                label: t('settingTheme.themeMode.dark'),
-                value: 'dark',
-              },
-              {
-                icon: Monitor,
-                img: imageUrl('theme_auto.webp'),
-                label: t('settingTheme.themeMode.auto'),
-                value: 'auto',
-              },
-            ]}
-            width={100}
-          />
-        ),
-        label: t('settingTheme.themeMode.title'),
-        minWidth: undefined,
-      },
-      {
-        children: (
-          <Select
-            onChange={switchLang}
-            options={[{ label: t('settingTheme.lang.autoMode'), value: 'auto' }, ...localeOptions]}
-          />
-        ),
-        label: t('settingTheme.lang.title'),
-        name: 'language',
-      },
-      {
-        children: (
-          <SliderWithInput
-            marks={{
-              12: {
-                label: 'A',
-                style: {
-                  fontSize: 12,
-                  marginTop: 4,
-                },
-              },
-              14: {
-                label: t('settingTheme.fontSize.marks.normal'),
-                style: {
-                  fontSize: 14,
-                  marginTop: 4,
-                },
-              },
-              18: {
-                label: 'A',
-                style: {
-                  fontSize: 18,
-                  marginTop: 4,
-                },
-              },
-            }}
-            max={18}
-            min={12}
-            step={1}
-          />
-        ),
-        desc: t('settingTheme.fontSize.desc'),
-        label: t('settingTheme.fontSize.title'),
-        name: 'fontSize',
-      },
-      {
-        children: <ThemeSwatchesPrimary />,
-        desc: t('settingTheme.primaryColor.desc'),
-        label: t('settingTheme.primaryColor.title'),
-        minWidth: undefined,
-      },
-      {
-        children: <ThemeSwatchesNeutral />,
-        desc: t('settingTheme.neutralColor.desc'),
-        label: t('settingTheme.neutralColor.title'),
-        minWidth: undefined,
-      },
-    ],
-    icon: Palette,
-    title: t('settingTheme.title'),
-  };
 
   const system: SettingItemGroup = {
     children: [
@@ -276,7 +164,7 @@ const Common = memo<SettingsCommonProps>(({ showAccessCodeConfig, showOAuthLogin
     <Form
       form={form}
       initialValues={settings}
-      items={[theme, system]}
+      items={[system]}
       onValuesChange={setSettings}
       {...FORM_STYLE}
     />
