+++ src/app/settings/common/index.tsx
@@ -8,6 +8,7 @@ import { CURRENT_VERSION } from '@/const/version';
 
 import Footer from '../features/Footer';
 import Common, { SettingsCommonProps } from './Common';
+import Theme from './Theme';
 
 export default memo<SettingsCommonProps>((props) => {
   const { t } = useTranslation('setting');
@@ -15,6 +16,7 @@ export default memo<SettingsCommonProps>((props) => {
   return (
     <>
       <PageTitle title={t('tab.common')} />
+      <Theme />
       <Common {...props} />
       <Footer>LobeChat v{CURRENT_VERSION}</Footer>
     </>
